bin/staticmatic: Fix lib_dir when staticmatic is a symlink.
If you symlink the staticmatic executable into e.g. /usr/bin, it will
use the symlink as starting point for the library path, which casues
staticmatic to not find its libraries.

This patch fixes that by using the real executable as starting point.
Also works if executable is not a symlink.